refactor(api/core/workflow/nodes/variable_assigner): Split into multi files.
