+++ api/core/workflow/nodes/variable_assigner/exc.py
@@ -1,0 +1,2 @@
+class VariableAssignerNodeError(Exception):
+    pass
